chore(backend): Extract authStatus related enum & types to separate module
This will make tokens/request module more readable
