+++ packages/backend/src/tokens/index.ts
@@ -1,11 +1,10 @@
 export * from './authObjects';
 export * from './factory';
+export { type RequestState, AuthStatus } from './authStatus';
 export { loadInterstitialFromLocal } from './interstitial';
 export {
-  AuthStatus,
   debugRequestState,
   type AuthenticateRequestOptions,
   type OptionalVerifyTokenOptions,
-  type RequestState,
   type RequiredVerifyTokenOptions,
 } from './request';
